vendor: Add vendor prefix to avoid naming colision
Google has prohibited the use of vendor/ODM for properties
without a vendor prefix. We have added the prefix for them
but left other entries untouched; it's time to change that.
Let's add a vendor prefix to all vendor-related entries to
avoid potential naming conflicts.
Services provided in the ODM partiton have not yet received
a prefix for backward compatibility with previous ODM
releases. Most likely, this will happen in the v-mr1 branch
in coordination with the ODM supplier before their initial
release.

Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>